Product offer

+++ src/modules/product/product.service.ts
@@ -16,6 +16,11 @@ import { ProductFilter } from './dto/filter/product.filter';
 import { Subcategory } from 'src/infrastructure/entities/category/subcategory.entity';
 import { ProductSubCategory } from 'src/infrastructure/entities/product/product-sub-category.entity';
 import { MostHitSubcategory } from 'src/infrastructure/entities/category/most-hit-subcategory.entity';
+import { ProductOffer } from 'src/infrastructure/entities/product/product-offer.entity';
+import { CreateProductOfferRequest } from './dto/request/create-product-offer.request';
+import { CategorySubCategory } from 'src/infrastructure/entities/category/category-subcategory.entity';
+import { Warehouse } from 'src/infrastructure/entities/warehouse/warehouse.entity';
+import { SingleProductRequest } from './dto/request/single-product.request';
 
 @Injectable()
 export class ProductService {
@@ -26,13 +31,20 @@ export class ProductService {
     private readonly productImageRepository: Repository<ProductImage>,
     @InjectRepository(ProductMeasurement)
     private readonly productMeasurementRepository: Repository<ProductMeasurement>,
-    @InjectRepository(Subcategory)
-    private subcategory_repo: Repository<Subcategory>,
+    @InjectRepository(CategorySubCategory)
+    private readonly categorySubcategory_repo: Repository<CategorySubCategory>,
     @InjectRepository(MostHitSubcategory)
-    private mostHitSubcategoryRepository: Repository<MostHitSubcategory>,
+    private readonly mostHitSubcategoryRepository: Repository<MostHitSubcategory>,
 
     @InjectRepository(ProductSubCategory)
-    private productSubCategory_repo: Repository<ProductSubCategory>,
+    private readonly productSubCategory_repo: Repository<ProductSubCategory>,
+
+    @InjectRepository(ProductOffer)
+    private productOffer_repo: Repository<ProductOffer>,
+
+    @InjectRepository(Warehouse)
+    private readonly warehouse_repo: Repository<Warehouse>,
+
     @Inject(CreateProductTransaction)
     private readonly addProductTransaction: CreateProductTransaction,
 
@@ -44,10 +56,23 @@ export class ProductService {
 
     @Inject(UpdateProductImageTransaction)
     private readonly updateProductImageTransaction: UpdateProductImageTransaction,
-  ) { }
+  ) {}
 
-  async create(createProductRequest: CreateProductRequest): Promise<Product> {
+  async createProduct(
+    createProductRequest: CreateProductRequest,
+  ): Promise<Product> {
     return await this.addProductTransaction.run(createProductRequest);
+  }
+
+  async createProductOffer(
+    product_category_price_id: string,
+    createProductOfferRequest: CreateProductOfferRequest,
+  ) {
+    const createProductOffer = this.productOffer_repo.create(
+      createProductOfferRequest,
+    );
+    createProductOffer.product_category_price_id = product_category_price_id;
+    return await this.productOffer_repo.save(createProductOffer);
   }
 
   async updateProduct(
@@ -91,54 +116,87 @@ export class ProductService {
 
   async subCategoryAllProducts(
     productFilter: ProductFilter,
-    sub_category_id: string,
+    categorySubCategory_id: string,
   ): Promise<Product[]> {
-    const { page, limit } = productFilter;
+    const { page, limit, userLatitude, userLongitude } = productFilter;
 
     const skip = (page - 1) * limit;
 
     //* Check if sub category exist
-    const subCategory = await this.subcategory_repo.findOne({
-      where: { id: sub_category_id },
+    const categorySubcategory = await this.categorySubcategory_repo.findOne({
+      where: { id: categorySubCategory_id },
     });
-    if (!subCategory) {
+    if (!categorySubcategory) {
       throw new NotFoundException(`Subcategory ID not found`);
     }
+    //*TODO: add hit
+    // const hitSubCategory = await this.mostHitSubcategoryRepository.findOne({
+    //   where: { sub_category_id },
+    // });
 
-    const hitSubCategory = await this.mostHitSubcategoryRepository.findOne({
-      where: { sub_category_id },
-    });
-
-    if (hitSubCategory) {
-      await this.mostHitSubcategoryRepository.update(
-        { sub_category_id },
-        { current_hit: hitSubCategory.current_hit + 1 },
-      );
-    } else {
-      const newSubCategory = this.mostHitSubcategoryRepository.create({
-        subcategory: subCategory,
-        current_hit: 1,
-      });
-      await this.mostHitSubcategoryRepository.save(newSubCategory);
+    // if (hitSubCategory) {
+    //   await this.mostHitSubcategoryRepository.update(
+    //     { sub_category_id },
+    //     { current_hit: hitSubCategory.current_hit + 1 },
+    //   );
+    // } else {
+    //   const newSubCategory = this.mostHitSubcategoryRepository.create({
+    //     subcategory: subCategory,
+    //     current_hit: 1,
+    //   });
+    //   await this.mostHitSubcategoryRepository.save(newSubCategory);
+    // }
+    let warehouses: Warehouse;
+    if (userLatitude && userLongitude) {
+      warehouses = await this.warehouse_repo
+        .createQueryBuilder('warehouse')
+        .orderBy(
+          `ST_Distance_Sphere(
+            ST_SRID(point(${userLatitude}, ${userLongitude}), 4326),
+            warehouse.location
+        )`,
+        )
+        .getOne();
     }
 
     return await this.productRepository.find({
       skip,
       take: limit,
+      // where: {
+      //   warehouses_products: {
+      //     warehouse_id: warehouses.id,
+      //   },
+      //   product_measurements: {
+      //     product_category_prices: {
+      //       product_sub_category: {
+      //         category_subCategory:{
+      //           section_category:{
+      //             section_id
+      //           }
+      //         }
+      //       },
+      //     },
+      //   },
+      // },
       where: {
-
+        warehouses_products: {
+          warehouse_id: warehouses?.id,
+        },
         product_measurements: {
           product_category_prices: {
             product_sub_category: {
-              sub_category_id,
+              categorySubCategory_id,
             },
           },
         },
       },
       relations: {
         product_images: true,
+        warehouses_products: true,
         product_measurements: {
           product_category_prices: {
+            product_offer: true,
+
             product_additional_services: {
               additional_service: true,
             },
@@ -151,25 +209,44 @@ export class ProductService {
 
   async singleProduct(
     product_id: string,
-    sub_category_id?: string,
+    singleProductRequest: SingleProductRequest,
   ): Promise<Product> {
+
+  
+    const { categorySubCategory_id, userLatitude, userLongitude } =
+      singleProductRequest;
     //* Check if sub category exist
-    if (sub_category_id) {
-      const subCategory = await this.subcategory_repo.findOne({
-        where: { id: sub_category_id },
+    if (categorySubCategory_id) {
+      const categorySubcategory = await this.categorySubcategory_repo.findOne({
+        where: { id: categorySubCategory_id },
       });
-      if (!subCategory) {
+      if (!categorySubcategory) {
         throw new NotFoundException(`Subcategory ID not found`);
       }
+    }
+    let warehouses: Warehouse;
+    if (userLatitude && userLongitude) {
+      warehouses = await this.warehouse_repo
+        .createQueryBuilder('warehouse')
+        .orderBy(
+          `ST_Distance_Sphere(
+            ST_SRID(point(${userLatitude}, ${userLongitude}), 4326),
+            warehouse.location
+        )`,
+        )
+        .getOne();
     }
 
     const product = await this.productRepository.findOne({
       where: {
         id: product_id,
+        warehouses_products: {
+          warehouse_id: warehouses?.id,
+        },
         product_measurements: {
           product_category_prices: {
             product_sub_category: {
-              sub_category_id,
+              categorySubCategory_id,
             },
           },
         },
@@ -178,6 +255,7 @@ export class ProductService {
         product_images: true,
         product_measurements: {
           product_category_prices: {
+            product_offer: true,
             product_additional_services: {
               additional_service: true,
             },
@@ -195,7 +273,12 @@ export class ProductService {
     product_id: string,
     image_id: string,
   ): Promise<ProductImage> {
-    await this.singleProduct(product_id);
+    const product = await this.productRepository.findOne({
+      where: { id: product_id },
+    });
+    if (!product) {
+      throw new NotFoundException('message_product_not_found');
+    }
 
     const productImage = await this.productImageRepository.findOne({
       where: { id: image_id },
@@ -209,8 +292,12 @@ export class ProductService {
     product_id: string,
     measurement_id: string,
   ): Promise<ProductMeasurement> {
-    await this.singleProduct(product_id);
-
+    const product = await this.productRepository.findOne({
+      where: { id: product_id },
+    });
+    if (!product) {
+      throw new NotFoundException('message_product_not_found');
+    }
     const productMeasurement = await this.productMeasurementRepository.findOne({
       where: { id: measurement_id },
     });
@@ -221,7 +308,12 @@ export class ProductService {
   }
 
   async deleteProduct(product_id: string): Promise<DeleteResult> {
-    await this.singleProduct(product_id);
+    const product = await this.productRepository.findOne({
+      where: { id: product_id },
+    });
+    if (!product) {
+      throw new NotFoundException('message_product_not_found');
+    }
     return await this.productRepository.delete({ id: product_id });
   }
 
@@ -229,7 +321,12 @@ export class ProductService {
     product_id: string,
     image_id: string,
   ): Promise<DeleteResult> {
-    const product = await this.singleProduct(product_id);
+    const product = await this.productRepository.findOne({
+      where: { id: product_id },
+    });
+    if (!product) {
+      throw new NotFoundException('message_product_not_found');
+    }
     if (product.product_images.length == 1) {
       throw new NotFoundException('There must be at least one photo');
     }
@@ -241,7 +338,12 @@ export class ProductService {
     product_id: string,
     measurement_id: string,
   ): Promise<DeleteResult> {
-    await this.singleProduct(product_id);
+    const product = await this.productRepository.findOne({
+      where: { id: product_id },
+    });
+    if (!product) {
+      throw new NotFoundException('message_product_not_found');
+    }
     const measurement = await this.SingleProductMeasurement(
       product_id,
       measurement_id,
